Remove align-items center body new divs add styles
Removed align-items: center on body, which solved the issue of the
section backgrounds not expanding to fill the screen.

Added new divs to make adjusting each sections size easier; this was
and easier adjustment than I thought it would be.

Meticulously went through and did minor adjustments on styles and flexes
of fonts, containers, and all elements to make it match the image as
much as possible.